close #12 : new sim object classification

+++ src/main/java/org/socialworld/objects/concrete/clothes/shirts/TestShirt.java
@@ -11,7 +11,7 @@ public class TestShirt extends Shirt {
 
 
 	@Override
-	public boolean checkObjectBelongsToGroup(short groupNumberSuffix) {
+	public boolean checkObjectBelongsToGroup(int groupNumberSuffix) {
 		if (groupNumberSuffix == GroupingOfSimulationObjects.GROUPING_NUMBER_SUFFIX_TEST) {
 			return true;
 		}
