+++ src/main/java/org/socialworld/objects/concrete/clothes/shoes/TestShoeLeft.java
@@ -20,7 +20,7 @@ public class TestShoeLeft extends Shoe {
 	}
 
 	@Override
-	public boolean checkObjectBelongsToGroup(short groupNumberSuffix) {
+	public boolean checkObjectBelongsToGroup(int groupNumberSuffix) {
 		if (groupNumberSuffix == GroupingOfSimulationObjects.GROUPING_NUMBER_SUFFIX_TEST) {
 			return true;
 		}
