+++ src/main/java/org/socialworld/objects/concrete/clothes/trousers/TestTrousers.java
@@ -11,7 +11,7 @@ public class TestTrousers extends Trousers {
 
 
 	@Override
-	public boolean checkObjectBelongsToGroup(short groupNumberSuffix) {
+	public boolean checkObjectBelongsToGroup(int groupNumberSuffix) {
 		if (groupNumberSuffix == GroupingOfSimulationObjects.GROUPING_NUMBER_SUFFIX_TEST) {
 			return true;
 		}
